+++ meongmory_iOS/Domain/Service/MapService.swift
@@ -12,14 +12,12 @@ import Alamofire
 class MapService {
     private let kakaoAPIManger = KakaoAPIManager()
     
-    func getAllLocation(x: String, y: String) {
-        MapKind.allCases.forEach { mapKind in
-            kakaoAPIManger.getLocalInfo(keyword: mapKind.getKeyword(), x: x, y: y)
+    func getSearchLocation(keyword: String, x: String, y: String, completion: @escaping ([Document]) -> Void) {
+        kakaoAPIManger.getLocalInfo(keyword: keyword, x: x, y: y) { response in
+            if let data = response.documents {
+                completion(data)
+            }
         }
-    }
-    
-    func getSearchLocation(keyword: String, x: String, y: String) {
-        
     }
